API/inputsapi/requirements.txt
Django====3.0.2
djangorestframework==3.11.0